chore: move room hooks near to room provider
