+++ src/hooks/chat.ts
@@ -1,5 +1,5 @@
 import { useSubscription } from 'react-stomp-hooks';
-import { useRoom } from '~/hooks/room';
+import { useRoom } from '~/providers/RoomProvider';
 import { chatMessage, ChatMessage, ChatRoom } from '~/types/chat';
 
 export const useChatMessages = ({
