+++ src/providers/RoomProvider.tsx
@@ -1,9 +1,11 @@
 'use client';
 
 import React, { PropsWithChildren } from 'react';
-import { createRoomStore, defaultInitState } from '~/stores/room';
+import { useStore } from 'zustand';
+import { createRoomStore, defaultInitState, RoomStore } from '~/stores/room';
 import { ChatMessage, ChatRoom } from '~/types/chat';
 import { Nullable } from '~/types/misc';
+import { assert } from '~/utils/assert';
 
 export type RoomStoreApi = ReturnType<typeof createRoomStore>;
 
@@ -41,4 +43,15 @@ export const RoomProvider: React.FC<PropsWithChildren<Props>> = ({
       {children}
     </RoomStoreContext.Provider>
   );
+};
+
+export const useRoomStore = <T,>(selector: (store: RoomStore) => T): T => {
+  const roomStoreContext = React.useContext(RoomStoreContext);
+  assert(roomStoreContext, 'useRoomStore must be used within <RoomProvider/>');
+
+  return useStore(roomStoreContext, selector);
+};
+
+export const useRoom = () => {
+  return useRoomStore((store) => store);
 };
